Add production index (#41)
* add update functionality and test

* fix import path issues

This took a while, but it should be a "proper python packge" which allows easier importing.

* Make overwriting the default.

Need this for overwriting the "latest" reports

* most of the way there

Still need to properly determine the production_week

* Don't need to worry about cache size on ubuntu runners

Originally put this in bc I forget where this was running. Since it is running on GitHub's ubunutu-latest, we don't need to keep the cache size small.

* pre-commit

* add prod_week

Use a LLM generated function to round up to the Friday >= today. Add tests that cover several weeks.

* pre-commit

* make sure we don't get in bad situations

* use correct blob path, and tell polars the schema

* add warning when doing prod run

* fix relative import

* bump to latest version

* explain how to run so that paths work as expected

* update naming schema, prod index location, docs

* update server entry point, and arg validation